Working toward CS caching in serial.
Simplifying streaming view logic to simply iterate until reps say they are done.
This is moving toward unification with Adaptive which does the same thing and has
more of the logic encoded in the rep/strat/us's than the view.

Started work on using appended cache, but it doesn't appear to work yet.

The parallel strats have not yet been touched.