Streaming builtin looks to be working.
I wasn't putting the 'on pipeline changed' code in the right place.
I am now asking for the pipeline to append on the first pass.
I fixed a caching bug related to resolution which is 0 not 1 if not present in datainfo.
I forgot to sort on cache priorities.
I was counting the append slot as a piece and thus missing the real first piece.
qt snafu prevented very first render from streaming fully. Since accept button had been lit it had to stop immediately.