harness now works
The test indicates that harness does what it should, allow piece control
and enable priority and meta info queries. There code is a bit of a hack
at this point but it does what it has to. For now I am going to move on
to other classes and hope to clean this up later.