Prioritization works.
Added hooks for harness to be associated with a piecelist and a piececache.
Compute pipeline priorities in the prioritizedstreamer, stream in order and terminate
early when possible. Test setup changed to have enough pieces so that some are skippable
and to process at full resolution. Now can see that once full set is run through, and
thus the meta info is read once, you don't have to render all pieces to see it.

Brought over piece and piecelist from client side caching branch.