First go at multires.
It appears to work to me because it renders through several wends
until done and then continues to render the whole domain at max
resolution. The Z clipping range makes it appear not to be working
on interaction, but that is a different aspect to fix.

Caching is not yet back in, so performance is bad.
I have not put in manual progression, reaping or coarsening.