Workaround for refine/reap non convergence problem.
Also I forgot to assign pcf to harness so reaps were not cleaning cache.
This approach of marking pieces as being reaped and resetting that flag
seems to work, but I have doubts about whether it fixes or just hides
the problem. I think I need a better termination condition for a real
fix.