Start on ParaView level infrastructure.
Fix VTK level config to work with PV level enabled.
Also rename it from Adaptive to Streaming since it is now generic.
Create stubs of the classes that I intend to use to work it into ParaView.
That itself may be useful as it is a starting point for new RenderView plugins.