Add streaming control to view and hook up all of the components.
The infrastructure is all here (at least for builtin), now I need to turn it
on and fix up problems. SMView now makes a StreamingDriver. SMRep now tells
PVRep what it's upstream PCF and Harness are. SMView tells PVView what it's
Driver is. Add Rep to View connects the harnesses to the driver.

I fixed a leak in VTK classes by making the internals private.