First streamed rendering, although I am not yet showing the last piece.
I hooked up render eventually to ParaView GUI. To do that I changed the
render eventually signature to give it an instance and made the SMView check
a Done flag that the callback sets in the PVView.

Finally I make the View mark all the representations Dirty when rendering is
not done. Otherwise, even if the pipeline updates, the render call doesn't
show the new data.

I also added options to use the other two streaming drivers in the XML.